+++ src/main/java/duke/Command.java
@@ -1,3 +1,5 @@
+package duke;
+
 public abstract class Command {
 
     public abstract void execute(Storage storage, TaskList tasks, Ui ui) throws DukeException;
