+++ src/main/java/duke/Duke.java
@@ -1,10 +1,4 @@
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Scanner;
+package duke;
 
 public class Duke {
