+++ src/main/java/duke/Event.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class Event extends Task {
 
     protected String at;
